Update network module
